+++ Ana_Deutlich.cs
@@ -64,7 +64,7 @@ namespace PachydermGH
             DA.GetData<int>(1, ref Dx);
 
             List<double> D = new List<double>();
-            foreach (float[] f in ETC.Value)
+            foreach (double[] f in ETC.Value)
             {
                 double[] s = new double[f.Length];
                 int start = 0;
